Moved into application file for serving via a WSGI framework
